refactor: NewProduct/useNewProduct 생성
