+++ src/refactoring/hooks/admin/useAdmin.ts
@@ -3,34 +3,6 @@ import { AdminActions, Coupon, Discount, Product } from '../../../types';
 import { updateProduct } from '../utils/adminUtils';
 
 export const useAdmin = (products: Product[], adminActions: AdminActions) => {
-  // 새 상품 추가
-  const [showNewProductForm, setShowNewProductForm] = useState(false);
-  const [newProduct, setNewProduct] = useState<Omit<Product, 'id'>>({
-    name: '',
-    price: 0,
-    stock: 0,
-    discounts: [],
-  });
-  const updateNewProduct = (
-    newProduct: Omit<Product, 'id'>,
-    targetKey: keyof Product,
-    newValue: Product[keyof Product],
-  ) => {
-    const updatedProduct = updateProduct(newProduct, targetKey, newValue);
-    setNewProduct(updatedProduct);
-  };
-  const handleAddNewProduct = () => {
-    const productWithId = { ...newProduct, id: Date.now().toString() };
-    adminActions.addProduct(productWithId);
-    setNewProduct({
-      name: '',
-      price: 0,
-      stock: 0,
-      discounts: [],
-    });
-    setShowNewProductForm(false);
-  };
-
   // 상품 상세 토글
   const [openProductIds, setOpenProductIds] = useState<Set<string>>(new Set());
   const toggleProductAccordion = (productId: string) => {
@@ -113,12 +85,6 @@ export const useAdmin = (products: Product[], adminActions: AdminActions) => {
   };
 
   return {
-    showNewProductForm,
-    setShowNewProductForm,
-    newProduct,
-    updateNewProduct,
-    handleAddNewProduct,
-
     openProductIds,
     toggleProductAccordion,
