+++ src/refactoring/pages/AdminPage.tsx
@@ -1,4 +1,5 @@
 import { AdminActions, Coupon, Product } from '../../types.ts';
+import { NewProduct } from '../components/admin/NewProduct.tsx';
 import { useAdmin } from '../hooks';
 
 interface Props {
@@ -17,12 +18,6 @@ export const AdminPage = ({ products, coupons, onProductUpdate, onProductAdd, on
   };
 
   const {
-    showNewProductForm,
-    setShowNewProductForm,
-    newProduct,
-    updateNewProduct,
-    handleAddNewProduct,
-
     openProductIds,
     toggleProductAccordion,
 
@@ -47,59 +42,9 @@ export const AdminPage = ({ products, coupons, onProductUpdate, onProductAdd, on
       <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
         <div>
           <h2 className="text-2xl font-semibold mb-4">상품 관리</h2>
-          <button
-            onClick={() => setShowNewProductForm(!showNewProductForm)}
-            className="bg-green-500 text-white px-4 py-2 rounded mb-4 hover:bg-green-600"
-          >
-            {showNewProductForm ? '취소' : '새 상품 추가'}
-          </button>
-          {showNewProductForm && (
-            <div className="bg-white p-4 rounded shadow mb-4">
-              <h3 className="text-xl font-semibold mb-2">새 상품 추가</h3>
-              <div className="mb-2">
-                <label htmlFor="productName" className="block text-sm font-medium text-gray-700">
-                  상품명
-                </label>
-                <input
-                  id="productName"
-                  type="text"
-                  value={newProduct.name}
-                  onChange={(e) => updateNewProduct(newProduct, 'name', e.target.value)}
-                  className="w-full p-2 border rounded"
-                />
-              </div>
-              <div className="mb-2">
-                <label htmlFor="productPrice" className="block text-sm font-medium text-gray-700">
-                  가격
-                </label>
-                <input
-                  id="productPrice"
-                  type="number"
-                  value={newProduct.price}
-                  onChange={(e) => updateNewProduct(newProduct, 'price', parseInt(e.target.value))}
-                  className="w-full p-2 border rounded"
-                />
-              </div>
-              <div className="mb-2">
-                <label htmlFor="productStock" className="block text-sm font-medium text-gray-700">
-                  재고
-                </label>
-                <input
-                  id="productStock"
-                  type="number"
-                  value={newProduct.stock}
-                  onChange={(e) => updateNewProduct(newProduct, 'stock', parseInt(e.target.value))}
-                  className="w-full p-2 border rounded"
-                />
-              </div>
-              <button
-                onClick={handleAddNewProduct}
-                className="w-full bg-blue-500 text-white p-2 rounded hover:bg-blue-600"
-              >
-                추가
-              </button>
-            </div>
-          )}
+
+          <NewProduct addProduct={onProductAdd} />
+
           <div className="space-y-2">
             {products.map((product, index) => (
               <div key={product.id} data-testid={`product-${index + 1}`} className="bg-white p-4 rounded shadow">
